# services/learning_api/openapi.json
{
  "components": {
    "schemas": {
...
      "ActiveRecallAttemptRequest": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "maxLength": 500,
            "title": "Answer",
            "type": "string"
          }
        },
        "required": [
          "answer"
        ],
        "title": "ActiveRecallAttemptRequest",
        "type": "object"
      },
      "ActiveRecallAttemptResponse": {
        "additionalProperties": false,
        "properties": {
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "feedback": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Feedback",
            "type": "array"
          },
          "progress": {
            "$ref": "#/components/schemas/LessonSectionProgressResponse"
          }
        },
        "required": [
          "correct",
          "error",
          "feedback",
          "progress"
        ],
        "title": "ActiveRecallAttemptResponse",
        "type": "object"
      },
...
      "CheckpointProgressResponse": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "last_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Percentage"
          },
          "last_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Session Id"
          },
          "passing_percentage": {
            "title": "Passing Percentage",
            "type": "integer"
          },
          "question_count": {
            "title": "Question Count",
            "type": "integer"
          },
          "state": {
            "enum": [
              "locked",
              "available",
              "in_progress",
              "passed"
            ],
            "title": "State",
            "type": "string"
          },
          "unit_key": {
            "title": "Unit Key",
            "type": "string"
          }
        },
        "required": [
          "unit_key",
          "state",
          "available",
          "question_count",
          "passing_percentage"
        ],
        "title": "CheckpointProgressResponse",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "LearningHomeResponse": {
        "additionalProperties": false,
        "properties": {
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/CheckpointProgressResponse"
            },
            "title": "Checkpoints",
            "type": "array"
          },
          "course_key": {
            "title": "Course Key",
            "type": "string"
          },
          "learner_id": {
            "title": "Learner Id",
            "type": "string"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonProgressResponse"
            },
            "title": "Lessons",
            "type": "array"
          },
          "next_action": {
            "$ref": "#/components/schemas/NextActionResponse"
          },
          "unresolved_retry_count": {
            "title": "Unresolved Retry Count",
            "type": "integer"
          }
        },
        "required": [
          "learner_id",
          "course_key",
          "next_action",
          "unresolved_retry_count",
          "lessons",
          "checkpoints"
        ],
        "title": "LearningHomeResponse",
        "type": "object"
      },
      "LessonProgressResponse": {
        "additionalProperties": false,
        "properties": {
          "checkpoint_passed": {
            "default": false,
            "title": "Checkpoint Passed",
            "type": "boolean"
          },
          "correct_count": {
            "default": 0,
            "title": "Correct Count",
            "type": "integer"
          },
          "eventual_correct_percentage": {
            "default": 0,
            "title": "Eventual Correct Percentage",
            "type": "number"
          },
          "gave_up_count": {
            "default": 0,
            "title": "Gave Up Count",
            "type": "integer"
          },
          "last_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Session Id"
          },
          "lesson_key": {
            "title": "Lesson Key",
            "type": "string"
          },
          "lesson_title": {
            "title": "Lesson Title",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "question_count": {
            "default": 0,
            "title": "Question Count",
            "type": "integer"
          },
          "resolved_count": {
            "default": 0,
            "title": "Resolved Count",
            "type": "integer"
          },
          "retry_question_count": {
            "default": 0,
            "title": "Retry Question Count",
            "type": "integer"
          },
          "state": {
            "enum": [
              "not_started",
              "in_progress",
              "practice_completed",
              "proficient",
              "mastered"
            ],
            "title": "State",
            "type": "string"
          },
          "unlock_reason": {
            "anyOf": [
              {
                "const": "prerequisite_not_proficient",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unlock Reason"
          },
          "unlocked": {
            "default": false,
            "title": "Unlocked",
            "type": "boolean"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "lesson_key",
          "lesson_title",
          "position",
          "state"
        ],
        "title": "LessonProgressResponse",
        "type": "object"
      },
...
      "LessonSectionProgressResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_count": {
            "minimum": 0.0,
            "title": "Completed Count",
            "type": "integer"
          },
          "completed_section_keys": {
            "items": {
              "type": "string"
            },
            "title": "Completed Section Keys",
            "type": "array"
          },
          "lesson_key": {
            "title": "Lesson Key",
            "type": "string"
          },
          "lesson_revision": {
            "minimum": 1.0,
            "title": "Lesson Revision",
            "type": "integer"
          },
          "total_sections": {
            "minimum": 0.0,
            "title": "Total Sections",
            "type": "integer"
          }
        },
        "required": [
          "lesson_key",
          "lesson_revision",
          "total_sections",
          "completed_count",
          "completed_section_keys"
        ],
        "title": "LessonSectionProgressResponse",
        "type": "object"
      },
...
      "ProgressResponse": {
        "additionalProperties": false,
        "properties": {
          "checkpoint_required_for_mastery": {
            "title": "Checkpoint Required For Mastery",
            "type": "boolean"
          },
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/CheckpointProgressResponse"
            },
            "title": "Checkpoints",
            "type": "array"
          },
          "course_key": {
            "title": "Course Key",
            "type": "string"
          },
          "learner_id": {
            "title": "Learner Id",
            "type": "string"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonProgressResponse"
            },
            "title": "Lessons",
            "type": "array"
          },
          "proficiency_threshold": {
            "title": "Proficiency Threshold",
            "type": "integer"
          }
        },
        "required": [
          "learner_id",
          "course_key",
          "proficiency_threshold",
          "checkpoint_required_for_mastery",
          "lessons",
          "checkpoints"
        ],
        "title": "ProgressResponse",
        "type": "object"
      },
...
      "SetLessonSectionCompletionRequest": {
        "additionalProperties": false,
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "boolean"
          }
        },
        "required": [
          "completed"
        ],
        "title": "SetLessonSectionCompletionRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned course and practice API. Lesson bodies remain draft placeholders until reviewed content and video assets are supplied.",
    "title": "ASEAN Academy Learning API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/learning-home": {
      "get": {
        "operationId": "getLearningHome",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningHomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get authenticated learner progress and the recommended next action",
        "tags": [
          "progress"
        ]
      }
    },
...
    "/api/v1/lessons/{lesson_key}/section-progress": {
      "get": {
        "operationId": "getLessonSectionProgress",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_key",
            "required": true,
            "schema": {
              "pattern": "^n1-lesson-[0-9]{2}$",
              "title": "Lesson Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonSectionProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get explicit section completion for the current lesson revision",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/v1/lessons/{lesson_key}/sections/{section_key}/active-recall": {
      "post": {
        "operationId": "checkLessonActiveRecall",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_key",
            "required": true,
            "schema": {
              "pattern": "^n1-lesson-[0-9]{2}$",
              "title": "Lesson Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "section_key",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9-]+$",
              "title": "Section Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveRecallAttemptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveRecallAttemptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Check an authored lesson recall answer without exposing its answer key",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/v1/lessons/{lesson_key}/sections/{section_key}/completion": {
      "put": {
        "operationId": "setLessonSectionCompletion",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_key",
            "required": true,
            "schema": {
              "pattern": "^n1-lesson-[0-9]{2}$",
              "title": "Lesson Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "section_key",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9-]+$",
              "title": "Section Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLessonSectionCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonSectionProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Explicitly mark a lesson note section complete or incomplete",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/v1/lessons/{lesson_key}/start": {
      "post": {
        "operationId": "startLesson",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_key",
            "required": true,
            "schema": {
              "pattern": "^n1-lesson-[0-9]{2}$",
              "title": "Lesson Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Idempotently start a lesson for the authenticated learner",
        "tags": [
          "progress"
        ]
      }
    },
...
    "/api/v1/progress": {
      "get": {
        "operationId": "getProgress",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get authenticated learner progress across the N1 course",
        "tags": [
          "progress"
        ]
      }
    },
...
  }
}